removed warnings
git-svn-id: http://mapsforge.googlecode.com/svn/trunk@296 a424f762-2dd8-11df-bfeb-7b4c664f286a

+++ mapsforge/src/org/mapsforge/preprocessing/map/Utils.java
@@ -281,8 +281,7 @@ class Utils {
 		}
 
 		for (Tile p : wayTiles) {
-			List<Tile> currentSubTiles = getSubtiles(new Tile((long) p.x, (long) p.y,
-					p.zoomLevel), (byte) 2);
+			List<Tile> currentSubTiles = getSubtiles(new Tile(p.x, p.y, p.zoomLevel), (byte) 2);
 			tileCounter = 0;
 			bitmap = 0;
 			result.put(p, (short) 0);
